requirements.txt
# ── Core ──────────────────────────────────────────────────────
aiogram>=3.0
fastapi
uvicorn
pydantic-settings
python-dotenv

# ── Logging ───────────────────────────────────────────────────
loguru

# ── AI ────────────────────────────────────────────────────────
openai
anthropic
google-generativeai

# ── RAG / Vector ──────────────────────────────────────────────
qdrant-client

# ── IoT ───────────────────────────────────────────────────────
aiomqtt
websockets

# ── Data & Viz ────────────────────────────────────────────────
numpy
pandas
matplotlib
plotly

# ── Crawler ───────────────────────────────────────────────────
playwright
selectolax
beautifulsoup4

# ── Utils ─────────────────────────────────────────────────────
httpx
apscheduler
redis

# ── UI ────────────────────────────────────────────────────────
rich

# ── Testing ───────────────────────────────────────────────────
pytest
pytest-asyncio
